rhel7 build fixes

+++ src/module/ihs/SGen_ihs/rechnen.f
@@ -401,7 +401,7 @@ C                                           A. Kaps  IHS,  01. 02. 99
 
       INCLUDE 'common.f'
 
-      REAL anzkugel
+      INTEGER anzkugel
       INTEGER i
 
       anzkugel=0
